refactor(oauth): remove the /oauth-external runtime surface
The dashboard playground was the sole consumer of the /oauth-external/*
endpoints (authorize, issuer-connect, status, disconnect, callback). With
the playground moving to the user-session paradigm (mint + first-party
connect), delete oauth/external_oauth.go, its hosted success-page assets,
the ExternalOAuthService wiring in start.go, and the now-dead
external-mcp-oauth-redirect-domains flag.

Relocate the symbols the main OAuth server still shares — the PKCE helpers
(generateCodeVerifier/generateCodeChallenge) into pkce.go and the DCR
request/response types into client_registration.go — so impl.go keeps
compiling. The toolset AddExternalOAuthServer CRUD and external_oauth DB
tables are left for the separate retirement effort.
