--- server/internal/oauth/hosted_external_oauth_success_script.js
@@ -1,3 +1,0 @@
-setTimeout(function () {
-  window.close();
-}, 1500);
