+++ server/internal/oauth/repo/models.go
@@ -9,21 +9,6 @@ import (
 	"github.com/jackc/pgx/v5/pgtype"
 )
 
-type ExternalOauthClientRegistration struct {
-	ID                    uuid.UUID
-	OrganizationID        string
-	ProjectID             uuid.UUID
-	OauthServerIssuer     string
-	ClientID              string
-	ClientSecretEncrypted pgtype.Text
-	ClientIDIssuedAt      pgtype.Timestamptz
-	ClientSecretExpiresAt pgtype.Timestamptz
-	CreatedAt             pgtype.Timestamptz
-	UpdatedAt             pgtype.Timestamptz
-	DeletedAt             pgtype.Timestamptz
-	Deleted               bool
-}
-
 type ExternalOauthServerMetadatum struct {
 	ID        uuid.UUID
 	ProjectID uuid.UUID
@@ -66,24 +51,4 @@ type OauthProxyServer struct {
 	UpdatedAt pgtype.Timestamptz
 	DeletedAt pgtype.Timestamptz
 	Deleted   bool
-}
-
-type UserOauthToken struct {
-	ID                    uuid.UUID
-	UserID                string
-	OrganizationID        string
-	ProjectID             uuid.UUID
-	ClientRegistrationID  uuid.UUID
-	ToolsetID             uuid.UUID
-	OauthServerIssuer     string
-	AccessTokenEncrypted  string
-	RefreshTokenEncrypted pgtype.Text
-	TokenType             pgtype.Text
-	ExpiresAt             pgtype.Timestamptz
-	Scopes                []string
-	ProviderName          pgtype.Text
-	CreatedAt             pgtype.Timestamptz
-	UpdatedAt             pgtype.Timestamptz
-	DeletedAt             pgtype.Timestamptz
-	Deleted               bool
 }
