+++ server/internal/oauth/repo/queries.sql
@@ -143,117 +143,3 @@ WHERE id = @id
   AND project_id = @project_id
   AND deleted IS FALSE
 RETURNING *;
-
--- User OAuth Tokens Queries
--- Stores tokens obtained from external OAuth providers for users authenticating to external MCP servers
-
--- name: UpsertUserOAuthToken :one
-INSERT INTO user_oauth_tokens (
-    user_id,
-    organization_id,
-    project_id,
-    client_registration_id,
-    toolset_id,
-    oauth_server_issuer,
-    access_token_encrypted,
-    refresh_token_encrypted,
-    token_type,
-    expires_at,
-    scopes,
-    provider_name
-) VALUES (
-    @user_id,
-    @organization_id,
-    @project_id,
-    @client_registration_id,
-    @toolset_id,
-    @oauth_server_issuer,
-    @access_token_encrypted,
-    @refresh_token_encrypted,
-    @token_type,
-    @expires_at,
-    @scopes,
-    @provider_name
-) ON CONFLICT (user_id, organization_id, toolset_id) WHERE deleted IS FALSE DO UPDATE SET
-    access_token_encrypted = EXCLUDED.access_token_encrypted,
-    refresh_token_encrypted = EXCLUDED.refresh_token_encrypted,
-    token_type = EXCLUDED.token_type,
-    expires_at = EXCLUDED.expires_at,
-    scopes = EXCLUDED.scopes,
-    provider_name = EXCLUDED.provider_name,
-    updated_at = clock_timestamp()
-RETURNING *;
-
--- name: GetUserOAuthToken :one
-SELECT * FROM user_oauth_tokens
-WHERE user_id = @user_id
-  AND organization_id = @organization_id
-  AND toolset_id = @toolset_id
-  AND deleted IS FALSE;
-
--- name: GetUserOAuthTokenByID :one
-SELECT * FROM user_oauth_tokens
-WHERE id = @id AND deleted IS FALSE;
-
--- name: ListUserOAuthTokens :many
-SELECT * FROM user_oauth_tokens
-WHERE user_id = @user_id
-  AND organization_id = @organization_id
-  AND deleted IS FALSE
-ORDER BY created_at DESC;
-
--- name: DeleteUserOAuthToken :exec
-UPDATE user_oauth_tokens SET
-    deleted_at = clock_timestamp(),
-    updated_at = clock_timestamp()
-WHERE id = @id;
-
--- name: DeleteUserOAuthTokenByToolset :exec
-UPDATE user_oauth_tokens SET
-    deleted_at = clock_timestamp(),
-    updated_at = clock_timestamp()
-WHERE user_id = @user_id
-  AND organization_id = @organization_id
-  AND toolset_id = @toolset_id;
-
--- External OAuth Client Registrations Queries
--- Stores client credentials from Dynamic Client Registration (DCR)
--- These are organization-level credentials, not user-level
-
--- name: UpsertExternalOAuthClientRegistration :one
-INSERT INTO external_oauth_client_registrations (
-    organization_id,
-    project_id,
-    oauth_server_issuer,
-    client_id,
-    client_secret_encrypted,
-    client_id_issued_at,
-    client_secret_expires_at
-) VALUES (
-    @organization_id,
-    @project_id,
-    @oauth_server_issuer,
-    @client_id,
-    @client_secret_encrypted,
-    @client_id_issued_at,
-    @client_secret_expires_at
-) ON CONFLICT (organization_id, oauth_server_issuer) WHERE deleted IS FALSE DO UPDATE SET
-    client_id = EXCLUDED.client_id,
-    client_secret_encrypted = EXCLUDED.client_secret_encrypted,
-    client_id_issued_at = EXCLUDED.client_id_issued_at,
-    client_secret_expires_at = EXCLUDED.client_secret_expires_at,
-    updated_at = clock_timestamp()
-RETURNING *;
-
--- name: GetExternalOAuthClientRegistration :one
-SELECT * FROM external_oauth_client_registrations
-WHERE organization_id = @organization_id
-  AND oauth_server_issuer = @oauth_server_issuer
-  AND deleted IS FALSE;
-
--- name: DeleteExternalOAuthClientRegistration :exec
-UPDATE external_oauth_client_registrations SET
-    deleted_at = clock_timestamp(),
-    updated_at = clock_timestamp()
-WHERE organization_id = @organization_id
-  AND oauth_server_issuer = @oauth_server_issuer;
